first

+++ src/fe-gtk/fe-gtk.h
@@ -171,8 +171,8 @@ typedef struct session_gui
 	int pane_left_size;	/*last position of the pane*/
 	int pane_right_size;
 
-	guint16 is_tab;	/* is tab or toplevel? */
-	guint16 ul_hidden;	/* userlist hidden? */
+	int is_tab;	/* is tab or toplevel? */
+	int ul_hidden;	/* userlist hidden? */
 
 } session_gui;
